pyproject.toml
[project]
name = "euler-preprocess"
version = "0.5.0"
description = "Physics-based preprocessing (fog, etc.) for RGB+depth datasets"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "Pillow",
  "euler-loading @ git+https://github.com/d-rothen/euler-loading.git",
]

[project.optional-dependencies]
gpu = [
  "torch",
]
progress = [
  "tqdm",
]

[project.scripts]
euler-preprocess = "euler_preprocess.cli:main"

[tool.setuptools.packages.find]
include = ["euler_preprocess", "euler_preprocess.*"]
